chore: Reorganize component files
